Move config.py in seperate package
Fixes bug that requires requests before being able to install
this package.
