+++ setup.py
@@ -2,7 +2,7 @@
 import codecs
 from os import path
 from setuptools import setup, find_packages
-from sap.xssec.config import USE_SAP_PY_JWT
+from sap.conf.config import USE_SAP_PY_JWT
 
 CURRENT_DIR = path.abspath(path.dirname(__file__))
 README_LOCATION = path.join(CURRENT_DIR, 'README.md')
@@ -28,7 +28,7 @@ setup(
     data_files=[('.', ['version.txt', 'CHANGELOG.md', 'sap_xssec.ini'])],
     test_suite='tests',
     install_requires=[
-        'requests>=2.21.0',
+        'requests>=2.22.0',
         'six==1.11.0',
         'pyjwt==1.7.*',
         '{}'.format(sap_py_jwt_dep)
